feat(fixed-deposits, kyc): Implement FD investment flow and KYC contract confirmation
This commit introduces the complete Fixed Deposit (FD) investment journey, including detailed product views, investment configuration, and purchase integration. Additionally, it implements the final KYC contract confirmation step with PDF preview capabilities.

**Key Changes:**

*   **Fixed Deposit (FD) Journey:**
    *   **FD Details Screen (`FDDetailsScreen.kt`):** A comprehensive view displaying bank info, risk levels, interest rate tables, lock-in periods, and FAQs. It supports interactive selection for interest payout modes and investment amounts.
    *   **FD Purchase Flow (`PurchaseFDScreen.kt`):** A dedicated screen for configuring investment details. Includes calculators for projected returns and maturity amounts based on selected tenures and frequencies.
    *   **ViewModels:** `FDDetailsViewModel` and `FDPurchaseViewModel` manage state for the respective screens, including data fetching and investment logic.
    *   **Data Layer:** Implements `FixedDepositRepo` and various DTOs/Mappers (`FDDetailsMapper`, `FdListMapper`) to handle FD search, details retrieval, and purchase URL generation.

*   **KYC Confirmation Flow:**
    *   **KYC Contract Screen (`KYCContractScreen.kt`):** A new screen for users to review their KYC document before finalization.
    *   **PDF Integration:** Introduces a `PdfViewer` interface with platform-specific implementations for Android (`Intent.ACTION_VIEW`) and iOS (`UIDocumentInteractionController`) to open KYC documents.
    *   **State Management:** `KycContractViewModel` handles fetching the contract PDF URL and finalizing the KYC process.
    *   **Use Cases:** Added `GetContractPdfUseCase` and `FinalizeKycUseCase`.

*   **User Data & Home Integration:**
    *   **User Data Mapper (`UserDataMapper.kt`):** Maps the unified `UserDataDto` to home screen UI models, including net worth, FIRE report progress, and KYC status.
    *   **Domain Models:** Updated to include `FDDetailsDomain`, `FDTenureDomain`, and user finance/goal structures.

*   **Utilities & Shared Components:**
    *   **HTML Parser (`HtmlToText.kt`):** A utility to convert HTML string content (from FAQs/About sections) into clean, readable text.
    *   **App Events:** Added `AppEvents.kt` using `SharedFlow` to handle cross-feature refresh events like `HomeEventRefresh`.
    *   **Calculators:** Logic for maturity amount and interest earned calculations.
    *   **UI Components:** New `PdfPreviewCard` and `ConsentSection` for the KYC flow.

+++ composeApp/src/commonMain/kotlin/org/sharad/velvetinvestment/data/remote/mapper/FireProjectionMapper.kt
@@ -19,8 +19,8 @@ fun FireReportDto.toDomain(): FireReportDomain {
             PortfolioProjectionPointDomain(
                 year = projection.year,
                 portfolioValue = EmiValueDomain(
-                    includeEmi = projection.portfolio_value.emi_include.toLong(),
-                    excludeEmi = projection.portfolio_value.emi_exclude.toLong()
+                    includeEmi = projection.portfolio_value.emi_include,
+                    excludeEmi = projection.portfolio_value.emi_exclude
                 )
             )
         )
@@ -38,25 +38,25 @@ fun FireReportDto.toDomain(): FireReportDomain {
         projectionRows.add(
             FireProjectionRowDomain(
                 year = projection.year,
-                income = projection.income.toLong(),
+                income = projection.income,
 
                 expenses = EmiValueDomain(
-                    includeEmi = projection.total_expenses.emi_include.toLong(),
-                    excludeEmi = projection.total_expenses.emi_exclude.toLong()
+                    includeEmi = projection.total_expenses.emi_include,
+                    excludeEmi = projection.total_expenses.emi_exclude
                 ),
 
-                yearlyGoalSip = projection.goal_commitment_annual.toLong(),
+                yearlyGoalSip = projection.goal_commitment_annual,
 
                 surplusMoney = EmiValueDomain(
-                    includeEmi = projection.savings.emi_include.toLong(),
-                    excludeEmi = projection.savings.emi_exclude.toLong()
+                    includeEmi = projection.savings.emi_include,
+                    excludeEmi = projection.savings.emi_exclude
                 ),
 
-                goalsFutureValue = projection.goals_payout.toLong(),
+                goalsFutureValue = projection.goals_payout,
 
                 portfolioValue = EmiValueDomain(
-                    includeEmi = projection.portfolio_value.emi_include.toLong(),
-                    excludeEmi = projection.portfolio_value.emi_exclude.toLong()
+                    includeEmi = projection.portfolio_value.emi_include,
+                    excludeEmi = projection.portfolio_value.emi_exclude
                 ),
 
                 fireNumber = EmiValueDomain(
